consumer1/requirements.txt
Flask==2.0.1
pika==1.2.0
werkzeug==2.0.2
requests
mysql-connector-python==8.0.28
